solana-perps-bot/requirements.txt
pandas==2.1.4
numpy==1.24.3
aiohttp==3.9.1
websockets==12.0
rich==13.7.0
loguru==0.7.2
python-dotenv==1.0.0
requests==2.31.0
